% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{army_in_the_us_deaths}
\alias{army_in_the_us_deaths}
\title{Death Due to Influenza in the United States Army
(in the United States)}
\format{
A tibble containing 13 rows and 2 columns.
\describe{
\item{\code{date}}{Date.}
\item{\code{deaths}}{Number of deaths.}
}
}
\source{
Howard, Deane C., and Love, Albert C., "Influenza-U.S. Army," Military Surgeon, vol.46 (May 1920),p. 525.
}
\usage{
army_in_the_us_deaths
}
\description{
Deaths Due to Influenza and Pneumonia between September 6th 1918
and November 29th 1918.
}
\keyword{datasets}
