% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deaths_by_sex_age}
\alias{deaths_by_sex_age}
\title{Number of Deaths from Influenza and Pneumonia by Sex and Age}
\format{
A tibble containing 112 rows and 5 columns
\describe{
\item{\code{age_of_decedent}}{Age of Decedent (in years).}
\item{\code{gender}}{Gender of Decedent.}
\item{\code{year}}{Year of death.}
\item{\code{number}}{Number of deaths in the United States (exclusive of Hawaii).}
\item{\code{borough}}{Distribution per 1,000 in the United States (exclusive of Hawaii).}
}
}
\source{
Bureau of Census, Mortality Statistics, 1919 (Washington, D.C.: Goverment
Printing Office, 1921), p. 30.
}
\usage{
deaths_by_sex_age
}
\description{
Number of Deaths from Influenza and Pneumonia, in the Registration Area,
by Sex and Age, and Distribution per 1,000, by Age of Descedent: 1918 and 1919
}
\keyword{datasets}
