% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deaths_estimated_vs_excess_nyc_boroughs}
\alias{deaths_estimated_vs_excess_nyc_boroughs}
\title{Estimated normal, actual, and excess number of deaths from
influenza and pneumonia in NYC Boroughs}
\format{
A tibble containing 450 rows and 6 columns
\describe{
\item{\code{borough}}{NYC borough.}
\item{\code{variable}}{Estimated normal number of deaths from influenza and pneumonia, actual deaths from influenza and pneumonia, and excess death from influenza and pneumonia.}
\item{\code{month}}{Last 4 Month or first 6 month.}
\item{\code{year}}{Year of death.}
\item{\code{deaths}}{Number of deaths.}
}
}
\source{
Bureau of Census, Mortality Statistics, 1919 (Washington, D.C.: Goverment
Printing Office, 1921), p. 30.
}
\usage{
deaths_estimated_vs_excess_nyc_boroughs
}
\description{
Estimated normal, actual, and excess number of deaths from
influenza and pneumonia in NYC Boroughs in
the last 4 month of 1918 and the first 6 month of 1919.
}
\keyword{datasets}
