% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deaths_registered_in_certain_cities}
\alias{deaths_registered_in_certain_cities}
\title{Death Registered as Due to Influenza and Pneumonia in Certain
Cities of the United States}
\format{
A tibble containing 1,249 rows and 9 columns
\describe{
\item{\code{city}}{City.}
\item{\code{state}}{State.}
\item{\code{population}}{Population (July 1, 1918 Estimated).}
\item{\code{date}}{Date of reporting (Week ended).}
\item{\code{first_case}}{Date when first case was recorded.}
\item{\code{mortality_acc}}{Mortality acceleration date.}
\item{\code{response_date}}{Date when first non-pharmaceutical interventions began.}
\item{\code{npi_speed_1918}}{Speed of non-pharmaceutical interventions in 1918.}
\item{\code{npi_days_1918}}{Days with non-pharmaceutical interventions in 1918.}
}
}
\source{
Great Britain, Ministry of Health, Reports on Public Health and Medical Subjects Number 4,
Report on the Pandemic of Influenza, 1918-19 (London: His Majesty's Stationery Office, 1920), pp. 319-20.

Markel, H., H. B. Lipman, J. A. Navarro, A. Sloan, J. R. Michalsen, A. M. Stern, and M. S.
Cetron (2007). Nonpharmaceutical Interventions Implemented by US Cities During the
1918-1919 Influenza Pandemic. JAMA 298(6), pp. 644–654.
}
\usage{
deaths_registered_in_certain_cities
}
\description{
Death Registered as Due to Influenza and Pneumonia in Certain
Cities of the United States, from September 8th 1918 to
March 15th 1919, by Weeks
}
\keyword{datasets}
