% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deaths_registered_in_certain_cities_1000}
\alias{deaths_registered_in_certain_cities_1000}
\title{Death from Influenza and Pneumonia per 1,000 Population
in Large U.S. Cities}
\format{
A tibble containing 250 rows and 9 columns
\describe{
\item{\code{city}}{City.}
\item{\code{state}}{State.}
\item{\code{year}}{Year.}
\item{\code{deaths_per_1000_population}}{Death per 1000 population.}
\item{\code{first_case}}{Date when first case was recorded.}
\item{\code{mortality_acc}}{Mortality acceleration date.}
\item{\code{response_date}}{Date when first non-pharmaceutical interventions began.}
\item{\code{npi_speed_1918}}{Speed of non-pharmaceutical interventions in 1918.}
\item{\code{npi_days_1918}}{Days with non-pharmaceutical interventions in 1918.}
}
}
\source{
Bureau of Census, Mortality Statistics, 1919 (Washington, D.C.: Government
Printing Office, 1921), pp. 28-29.

#' Markel, H., H. B. Lipman, J. A. Navarro, A. Sloan, J. R. Michalsen, A. M. Stern, and M. S.
Cetron (2007). Nonpharmaceutical Interventions Implemented by US Cities During the
1918-1919 Influenza Pandemic. JAMA 298(6), pp. 644–654.
}
\usage{
deaths_registered_in_certain_cities_1000
}
\description{
Death from Influenza and Pneumonia per 1,000 Population
in Large U.S. Cities between 1915 and 1919 (per year)
}
\keyword{datasets}
