% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deaths_registered_in_certain_states_1000}
\alias{deaths_registered_in_certain_states_1000}
\title{Death from Influenza and Pneumonia per 1,000 Population
in U.S. States}
\format{
A tibble containing 120 rows and 3 columns
\describe{
\item{\code{state}}{State.}
\item{\code{year}}{Year.}
\item{\code{deaths_per_1000_population}}{Death per 1000 population.}
}
}
\source{
Bureau of Census, Mortality Statistics, 1919 (Washington, D.C.: Government
Printing Office, 1921), pp. 28-29.
}
\usage{
deaths_registered_in_certain_states_1000
}
\description{
Death from Influenza and Pneumonia per 1,000 Population
in U.S. States between 1915 and 1919 (per year)
}
\keyword{datasets}
