% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naval_forces_in_the_us_deaths}
\alias{naval_forces_in_the_us_deaths}
\title{Death Due to Influenza and Pneumonia in the United States Navel Forces
(Ashore in the United States)}
\format{
A tibble containing 13 rows and 2 columns.
\describe{
\item{\code{date}}{Date.}
\item{\code{deaths}}{Number of deaths.}
}
}
\source{
Annual Reports of the Navy Department, 1919, p. 2448.
}
\usage{
naval_forces_in_the_us_deaths
}
\description{
Deaths Due to Influenza and Pneumonia between August 31th 1918
and November 30th 1918.
}
\keyword{datasets}
